% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_raw_mac}
\alias{as_raw_mac}
\title{Convert a charactrer vector of MAC addresses to
a \code{list} of \code{raw} vectors.}
\usage{
as_raw_mac(mac)
}
\arguments{
\item{mac}{character vector of MAC address strings}
}
\description{
Convert a charactrer vector of MAC addresses to
a \code{list} of \code{raw} vectors.
}
\examples{
as_raw_mac("c8:69:cd:28:5a:7d")
}
