% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_canonical_mac}
\alias{is_canonical_mac}
\title{Test if MAC address strings are in canonical form}
\usage{
is_canonical_mac(mac)
}
\arguments{
\item{mac}{character vector of MAC address strings}
}
\description{
The canonical form for MAC addresses strings is 6, 2-character
octets separated by colons. This function returns \code{TRUE} if
an address string is in canonical format.
}
\examples{
is_canonical_mac("c8:69:cd:28:5a:7d")
}
