% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mac_age_db}
\alias{mac_age_db}
\title{MAC Age Database}
\description{
For now just see <https://github.com/hdm/mac-ages> and use [update_mac_age_db()] to update it.
}
\references{
<https://github.com/hdm/mac-ages>
}
\keyword{datasets}
