% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mac_registry_data}
\alias{mac_registry_data}
\title{MAC Registry Data}
\description{
For now just see <https://github.com/hdm/mac-ages> and use [update_registry_data()] to update it.
}
\references{
<https://github.com/hdm/mac-ages>
}
\keyword{datasets}
