% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac-to-binary-string.R
\name{mac_to_binary_string}
\alias{mac_to_binary_string}
\title{Convert MAC address character vector to a binary string representation}
\usage{
mac_to_binary_string(x)
}
\arguments{
\item{x}{character vector of MAC address. Each MAC address can be in any
case, the octets do not need to be 0-prefixed, and it doesn't matter
if the octets are separated by a \code{:} or if they are just contiguous.
If they _are_contiguous, that — by definition — means
each octet is 0-prefixed.}
}
\value{
a character vector of binary strings
}
\description{
Convert MAC address character vector to a binary string representation
}
\examples{
mac_to_binary_string(c("f023b9eb4204", "f0:23:b9:eb:42:4", "F023B9eB4204"))
}
