% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACtools-package.R
\docType{package}
\name{MACtools}
\alias{MACtools}
\alias{MACtools-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/MACtools}
\item BugReports: \url{https://gitlab.com/hrbrmstr/MACtools/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
