% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{rebuild_search_tries}
\alias{rebuild_search_tries}
\title{Rebuild in-memory search tries}
\usage{
rebuild_search_tries()
}
\description{
The search structures created by \code{triebeard} are full of external pointers
which point to the deep, dark void when bad things happen to R sessions.
When the package reloads everything \emph{should} return to normal, but you
can use this function to rebuild the in-memory search structures at any time.
}
