% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACtools-package.R
\docType{package}
\name{MACtools}
\alias{MACtools}
\alias{MACtools-package}
\title{Tools to Work with Media Access Control ('MAC') Addresses}
\description{
A media access control address (MAC address) of a device is a unique
identifier assigned to a network interface controller (NIC) for
communications at the data link layer of a network segment. MAC addresses are
used as a network address for most IEEE 802 network technologies, including
Ethernet and Wi-Fi. In this context, MAC addresses are used in the medium
access control protocol sublayer. Tools are provided to work with these 'MAC' addresses.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/MACtools}
\item BugReports: \url{https://gitlab.com/hrbrmstr/MACtools/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
